Move stack trace test extension into fixtures
